updated name for consistancy and added tag for SchoolHome.feature.
